+++ tests/test_convert.py
@@ -77,7 +77,7 @@ def test_convert_output(fresh_db, drop, expected):
 
 def test_convert_output_multiple_column_error(fresh_db):
     table = fresh_db["table"]
-    with pytest.raises(AssertionError) as excinfo:
+    with pytest.raises(ValueError) as excinfo:
         table.convert(["title", "other"], lambda v: v, output="out")
         assert "output= can only be used with a single column" in str(excinfo.value)